moved out cardselector
