+++ frontend/src/components/CardSelector.module.css
@@ -1,0 +1,23 @@
+.btnContainer {
+  display: flex;
+  padding: 16px;
+  justify-content: space-evenly;
+  flex-wrap: wrap;
+}
+
+.btnsDisabled {
+  pointer-events: none;
+  opacity: 0.3;
+}
+
+.card {
+  width: 50px;
+  padding: 4px;
+
+  cursor: pointer;
+}
+
+.selectedCard {
+  width: 60px;
+  padding: 4px;
+}
